Staff bugfix, SQL Explorer, Jupyter (#11)
### Changes
Speed up with `uv pip`, including moving packages from
`.pre-commit-config.yaml` to `requirements.{in,txt}` or `package.json`.

Upgrade to Python 3.12.

Use the JWKSClient instead of fetching and caching keys ourselves.

Add
[django-sql-explorer](https://github.com/chrisclark/django-sql-explorer)
to demodj and add a `login_view_wrapper()` for it.

Move code around so the basic functionality can be reused across the
Django admin site, Django Rest Framework (DRF), and JupyterLab
(Tornado), and eventually Flask Application Builder (FAB) for Airflow
too.

Call `user.save()` after granting staff status.

### Potential followup
* More unit tests for both Python and Typescript, and improving them to
not hit the network
* Two end-to-end tests: access denied and access granted
* Eventually some integration tests
* Explore PyJWT further, including module-level JWKClient and caching
behavior. For example, if a key is removed from the JWKS response,
should it stay in the cache or be removed?